+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/LastAccessReducer.kt
@@ -6,6 +6,7 @@ package mozilla.components.browser.state.reducer
 
 import mozilla.components.browser.state.action.LastAccessAction
 import mozilla.components.browser.state.action.LastAccessAction.UpdateLastAccessAction
+import mozilla.components.browser.state.action.LastAccessAction.UpdateLastMediaAccessAction
 import mozilla.components.browser.state.state.BrowserState
 import mozilla.components.browser.state.state.TabSessionState
 
@@ -19,6 +20,12 @@ internal object LastAccessReducer {
             state.updateTabState(action.tabId) { sessionState ->
                 val tabSessionState = sessionState as TabSessionState
                 tabSessionState.copy(lastAccess = action.lastAccess)
+            }
+        }
+        is UpdateLastMediaAccessAction -> {
+            state.updateTabState(action.tabId) { sessionState ->
+                val tabSessionState = sessionState as TabSessionState
+                tabSessionState.copy(lastMediaAccess = action.lastMediaAccess)
             }
         }
     }